Rename generateemail package
